requirements.txt
absl-py==1.4.0
array_record==0.8.1
asgiref==3.9.1
astunparse==1.6.3
attrs==25.3.0
bleach==6.2.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Cython==3.1.4
Django==5.2.6
django-cors-headers==4.8.0
django-tastypie==0.15.1
djangorestframework==3.16.1
dm-tree==0.1.9
docstring_parser==0.17.0
einops==0.8.1
etils==1.13.0
flatbuffers==25.2.10
fonttools==4.59.2
fsspec==2025.9.0
gast==0.4.0
gin-config==0.5.0
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.74.0
gunicorn==23.0.0
h5py==3.14.0
httplib2==0.31.0
idna==3.10
immutabledict==4.2.1
importlib_resources==6.5.2
joblib==1.5.2
kaggle==1.7.4.5
keras==2.14.0
kiwisolver==1.4.9
libclang==18.1.1
lxml==6.0.1
Markdown==3.9
MarkupSafe==3.0.2
matplotlib==3.10.6
ml-dtypes==0.2.0
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.3.1
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
portalocker==3.2.0
promise==2.3
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.0.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.10
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-mimeparse==2.0.0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
sacrebleu==2.5.1
scikit-learn==1.7.2
scipy==1.16.2
sentencepiece==0.2.1
seqeval==1.2.2
simple-parsing==0.1.7
six==1.17.0
sqlparse==0.5.3
tabulate==0.9.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.1
tensorflow-datasets==4.9.9
tensorflow-estimator==2.14.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.17.2
tensorflow-model-optimization==0.8.0
tensorflow-text==2.14.0
termcolor==3.1.0
text-unidecode==1.3
tf-keras==2.15.0
tf-models-official==2.14.2
tf-slim==1.1.0
threadpoolctl==3.6.0
toml==0.10.2
tqdm==4.67.1
typing_extensions==4.5.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.14.2
zipp==3.23.0
